backend/requirements.txt
fastapi
uvicorn
paho-mqtt
websockets
python-dotenv
SQLAlchemy
asyncpg
pydantic